refactor: consolidate import and manifest callbacks into DDL callback framework
Move import callbacks and batch update manifest callback into DDLCallbacks
struct following the established pattern (e.g., ddl_callbacks_flushall.go).
Remove redundant initMessageCallback and drop partition callback registration.
