+++ internal/datacoord/ddl_callbacks_batch_update_manifest.go
@@ -1,0 +1,42 @@
+// Licensed to the LF AI & Data foundation under one
+// or more contributor license agreements. See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership. The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License. You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package datacoord
+
+import (
+	"context"
+
+	"go.uber.org/zap"
+
+	"github.com/milvus-io/milvus/pkg/v2/log"
+	"github.com/milvus-io/milvus/pkg/v2/streaming/util/message"
+)
+
+func (c *DDLCallbacks) batchUpdateManifestV2AckCallback(ctx context.Context, result message.BroadcastResultBatchUpdateManifestMessageV2) error {
+	body := result.Message.MustBody()
+	var operators []UpdateOperator
+	for _, item := range body.GetItems() {
+		operators = append(operators, UpdateManifestVersion(item.GetSegmentId(), item.GetManifestVersion()))
+	}
+	if len(operators) > 0 {
+		if err := c.meta.UpdateSegmentsInfo(ctx, operators...); err != nil {
+			log.Ctx(ctx).Warn("batch update manifest version failed", zap.Error(err))
+			return err
+		}
+	}
+	log.Ctx(ctx).Info("batch update manifest version handled", zap.Int("itemCount", len(body.GetItems())))
+	return nil
+}
